refactor: extract LogSessionForm component from erg-dashboard.jsx
Moves 399-line form component to web/src/components/LogSessionForm.jsx.
All component extractions complete — monolith now ~1,100 lines lighter.

Build and 68 tests pass.
